Use pattern_key as a non-unique pandas index for missingness data
Previously, Missingness used the row number of the original data as
the (unique) index. Using the new method, several operations can be
implemented more directly, and there is an overall performance
benefit (see below).

There is a performance (complexity) cost associated with using a
non-unique key in Pandas (see
e.g. https://stackoverflow.com/questions/16626058/what-is-the-performance-impact-of-non-unique-indexes-in-pandas),
which is improved by sorting on construction.  Using the previous
implementation, there were similar costs for the operations of
interest, just not associated with indexing.

  Performance comparison, using test_data_DIAG_10000000.csv

    pandas.read_csv (both): 27 s

    Missingness.from_data_frame (old): 9.7 s
    Missingness.from_data_frame (new): 9.9 s

    Missingness.counts (old): 126 ms
    Missingness.counts (new):  46 ms

    Missingness.matches (old): 38 ms
    Missingness.matches (new): 20 ms

    Missingness.select_columns (old): 580 ms
    Missingness.select_columns (new): 320 ms

    value_bar_chart_data (old): 1.7 s
    value_bar_chart_data (new): 0.6 s